@@ -4,30 +4,30 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": "7.4.*",
-        "ext-ctype": "*",
-        "ext-iconv": "*",
-        "composer/package-versions-deprecated": "1.11.99.2",
-        "doctrine/doctrine-bundle": "^2.4",
-        "doctrine/doctrine-migrations-bundle": "^3.1",
-        "doctrine/orm": "^2.9",
-        "symfony/console": "5.3.*",
-        "symfony/dotenv": "5.3.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "5.3.*",
-        "symfony/proxy-manager-bridge": "5.3.*",
-        "symfony/runtime": "5.3.*",
-        "symfony/serializer": "5.3.*",
-        "symfony/yaml": "5.3.*",
-        "thruway/pawl-transport": "^0.5.1",
-        "voryx/thruway": "^0.6.0"
+        "php": ">=8.2",
+        "doctrine/dbal": "^3",
+        "doctrine/doctrine-bundle": "^2.12",
+        "doctrine/doctrine-migrations-bundle": "^3.3",
+        "doctrine/orm": "^3.2",
+        "symfony/console": "6.3.*",
+        "symfony/dotenv": "6.3.*",
+        "symfony/flex": "^2.4.5",
+        "symfony/runtime": "6.3.*",
+        "symfony/serializer": "6.3.*",
+        "symfony/yaml": "6.3.*",
+        "thruway/pawl-transport": "^0.5.2",
+        "voryx/thruway": "^0.6.1"
     },
     "config": {
         "optimize-autoloader": true,
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "symfony/flex": true,
+            "symfony/runtime": true
+        }
     },
     "autoload": {
         "psr-4": {
@@ -62,7 +62,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.3.*"
+            "require": "6.3.*"
         }
     }
 }
