@@ -10,7 +10,7 @@ use Doctrine\Migrations\AbstractMigration;
 /**
  * Auto-generated Migration: Please modify to your needs!
  */
-final class Version20211006155636 extends AbstractMigration
+final class Version20240612145841 extends AbstractMigration
 {
     public function getDescription(): string
     {
