@@ -5,7 +5,9 @@ namespace App\Command\Wamp;
 use App\WampServer\WampRouter;
 use React\EventLoop\Loop;
 use React\EventLoop\LoopInterface;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Thruway\ClientSession;
@@ -13,32 +15,19 @@ use Thruway\Peer\Router;
 use Thruway\Transport\PawlTransportProvider;
 use Thruway\Transport\RatchetTransportProvider;
 
+#[AsCommand('app:start')]
 class StartWampServerCommand extends Command
 {
-    protected static $defaultName = "app:start";
-
-    protected WampRouter $wampRouter;
-
-    protected LoopInterface $loop;
-
-    public function __construct()
+    public function configure(): void
     {
-        parent::__construct();
-        $this->loop = Loop::get();
-    }
-
-    /** @required */
-    public function setWampRouter(WampRouter $wampRouter): self
-    {
-        $this->wampRouter = $wampRouter;
-
-        return $this;
+        $this->addArgument('server-url', InputArgument::REQUIRED, 'Url of the websocket server');
     }
 
     public function execute(InputInterface $input, OutputInterface $output): int
     {
-        $this->wampRouter->start(false);
-        $this->loop->run();
+        $wampRouter = new WampRouter($input->getArgument('server-url'));
+        $wampRouter->start(false);
+        Loop::get()->run();
 
         return Command::SUCCESS;
     }
